move a few classes
