+++ DDDEastAnglia.Tests/Admin/UserControllerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Net;
 using DDDEastAnglia.Areas.Admin.Controllers;
 using DDDEastAnglia.DataAccess;
+using DDDEastAnglia.Helpers;
 using DDDEastAnglia.Models;
 using NSubstitute;
 using NUnit.Framework;
