+++ DDDEastAnglia/DDDEastAnglia.csproj
@@ -235,8 +235,8 @@
     <Compile Include="DataAccess\ConferenceLoader.cs" />
     <Compile Include="DataAccess\IAccountLoginMethodQuery.cs" />
     <Compile Include="DataAccess\ISpeakerRepository.cs" />
-    <Compile Include="LoginMethodLoader.cs" />
-    <Compile Include="LoginMethodViewModelBuilder.cs" />
+    <Compile Include="Helpers\LoginMethodLoader.cs" />
+    <Compile Include="Helpers\LoginMethodViewModelBuilder.cs" />
     <Compile Include="Models\LoginMethod.cs" />
     <Compile Include="DataAccess\SimpleData\Queries\AccountLoginMethodQuery.cs" />
     <Compile Include="DataAccess\SpeakerRepository.cs" />
